+++ CHANGES.rst
@@ -4,7 +4,16 @@ Changes
 1.1a3 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- ``node_info`` decorator stores name on node info instance.
+  [rnix]
+
+- ``node_info`` decorator accepts additional keyword arguments to add custom
+  properties.
+  [rnix]
+
+- Add ``node_available`` callback to check whether node is allowed to be
+  used in application.
+  [rnix]
 
 
 1.1a2 (2024-02-12)